+++ packages/scripts/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Enhancements
 
--   `lint-style`: Detect .cjs and .mjs config files so the bundled default config is not used when one of these is present.
+-   `lint-style`: Detect .cjs and .mjs config files so the bundled default config is not used when one of these is present ([#79226](https://github.com/WordPress/gutenberg/pull/79226)).
 
 ## 32.4.0 (2026-06-10)